pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airfusion-ai"
version = "0.1.0"
description = "Multi-scale air quality data fusion and AI analysis service."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "openai>=1.99.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "paho-mqtt>=2.1.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
